Version 0.74
This release completely eliminates command-based snippets in favor of a new, easier to read & maintain syntax. A python script, convertSnip.py, has been included in this release to make it easier to migrate. For more information, see :h snippet-syntax; it's not very complicated, so it shouldn't take too much time to understand.

This release also fixes a number of bugs, including: support for supertab, special characters in snippet variables, the popup-menu, and adds support for hyphened snippet folders (e.g. snippets/c-snippets/ for C).